pyproject.toml
[project]
name = "iot-sample"
dynamic = ["version"]
requires-python = ">= 3.12"

dependencies = [
    "protobuf>=7.34.0",
    "pydantic-settings>=2.4.0",
    "google>=3.0.0",
    "fastapi>=0.112.0",
    "confluent-kafka>=2.5.0",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "uvicorn[standard]>=0.30.6",
    "faker>=28.4.1",
    "websockets>=13.0.1",
    "iot-proto",
]

[dependency-groups]
dev = [
    "python-semantic-release>=9.14.0",
    "ruff>=0.12.0",
    "fastapi[standard]>=0.112.0",
    "pre-commit>=4.2.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ty>=0.0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/iot_sample"]

[tool.hatch.version]
path = "src/iot_sample/__init__.py"

# Specifies a package that is a workspace member; this is applied to all members
# of the workspace
[tool.uv.sources]
iot-proto = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]
exclude = ["packages/iot-pki"]


[tool.ruff]
lint.select = ["E4", "E7", "E9", "F", "I"]
lint.ignore = []

line-length = 88
indent-width = 4

target-version = "py312"

# Exclude a variety of commonly ignored directories.
exclude = [
    "*_pb2.py",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

[tool.ruff.lint.isort]
known-first-party = ["iot_proto"]

[tool.semantic_release]
version_variables = [
    "src/iot_sample/__init__.py:__version__",
    "packages/iot-proto/python/__init__.py:__version__",
]
commit_author = {env = "GIT_COMMIT_AUTHOR"}

[tool.semantic_release.changelog]

[tool.semantic_release.publish]
upload_to_vcs_release = false

[tool.semantic_release.remote]
name = "gitea"
type = "gitea"

[tool.pytest]
minversion = "9.0"
asyncio_mode = "auto"
testpaths = ["src/python_tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
